make gruid.Grid Cells Color and AttrMask always uint32
The change should be backward compatible in practice, because depending
on the exact size of uint would have been an error, and 3*32 bits should
be enough to encode any generic color and attribute information that is
then mapped to drivers anyways. This change makes things clear, and
saves some space too, in particular because the Cell struct contains a
rune that is int32, so there was some padding.